admin/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.5.0
django-mysql==3.9.0
djangorestframework==3.12.4
mysqlclient==2.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
